% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IUCNhull2poly.R
\name{ahull.IUCN2lines}
\alias{ahull.IUCN2lines}
\title{Convert ahull.IUCN to a SpatialLines or sf LINESTRING object.}
\usage{
ahull.IUCN2lines(iucn_hull, sp_or_sf)
}
\arguments{
\item{iucn_hull}{The ahull.IUCN object}

\item{sp_or_sf}{Either "sp" or "sf" depending on the type of object you want to return.}
}
\value{
Either an object of class "SpatialLines", if [[sp_or_sf]] is "sp" or of class "sf" if [[sp_or_sf]] is "sf".
}
\description{
Convert ahull.IUCN to a SpatialLines or sf LINESTRING object.
}
