% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hull2poly.R
\name{ahull2poly}
\alias{ahull2poly}
\title{Convert alpha hull to SpatialPolygon}
\usage{
ahull2poly(hull)
}
\arguments{
\item{hull}{The alpha hull object}
}
\value{
The SpatialPolygon object created by that alpha hull
}
\description{
Convert alpha hull to SpatialPolygon
}
\examples{
data(iris)
iris_sepals <- iris[,1:2]
iris_sepals <- iris_sepals[!duplicated(paste(iris_sepals$Sepal.Length, iris_sepals$Sepal.Width)), ]
alphahull_1 <- alphahull::ahull(iris_sepals, alpha = 1)
hullpoly_1 <- ahull2poly(alphahull_1)
plot(iris_sepals, pch = 19, col = "darkseagreen")
plot(alphahull_1, lwd = 5, col = "gray", add = TRUE)
sp:::plot.SpatialPolygons(hullpoly_1, border = "magenta", add = TRUE)
}
