% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphashape2poly.R
\name{ashape2poly}
\alias{ashape2poly}
\title{Convert an alpha shape object to a polygon.}
\usage{
ashape2poly(ashape)
}
\arguments{
\item{ashape}{An alpha shape object.}
}
\value{
The polygon matching the boundaries of the alpha shape.
}
\description{
Convert an alpha shape object to a polygon.
}
\examples{
add(1, 1)
add(10, 1)
}
