% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hull2poly.R
\name{spLines2poly}
\alias{spLines2poly}
\title{Convert SpatialLines to SpatialPolygon}
\usage{
spLines2poly(sp_lines)
}
\arguments{
\item{sp_lines}{The SpatialLines object}
}
\value{
The SpatialPolygon object created by those lines
}
\description{
Convert SpatialLines to SpatialPolygon
}
