% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hull2poly.R
\name{ahull2poly}
\alias{ahull2poly}
\title{Convert alpha hull to SpatialPolygon}
\usage{
ahull2poly(hull)
}
\arguments{
\item{hull}{The alpha hull object}
}
\value{
The SpatialPolygon object created by that alpha hull
}
\description{
Convert alpha hull to SpatialPolygon
}
\examples{
add(1, 1)
add(10, 1)
}
