% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hull2line.R
\name{arc2line}
\alias{arc2line}
\title{Convert an arc into line segments given the center of the arc, the radius, the vector, and the angle (radians)}
\usage{
arc2line(center, r, vector, theta, npoints = 100)
}
\arguments{
\item{center}{The coordinates of the center}

\item{r}{The radius of the arc}

\item{vector}{The vector of the arc}

\item{theta}{The angle of the arc in radians}

\item{npoints}{Number of points along the arc}
}
\value{
The Line object that approximates the given arc
}
\description{
Convert an arc into line segments given the center of the arc, the radius, the vector, and the angle (radians)
}
\examples{
add(1, 1)
add(10, 1)
}
