% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphashape2poly.R
\name{ashape2poly}
\alias{ashape2poly}
\title{Convert an alpha shape object to a polygon. Modified from https://rpubs.com/geospacedman/alphasimple.}
\usage{
ashape2poly(ashape)
}
\arguments{
\item{ashape}{An alpha shape object.}
}
\value{
The polygon matching the boundaries of the alpha shape.
}
\description{
Convert an alpha shape object to a polygon. Modified from https://rpubs.com/geospacedman/alphasimple.
}
\examples{
set.seed(123)
x <- matrix(runif(100), nc = 2)
ashape_02 <- alphahull::ashape(x, alpha = 0.2)
poly_result <- ashape2poly(ashape_02)
}
